refactor: extract token utils to prevent test side effects
Moves flatten() and resolveRefs() into token-utils.ts so tests can
import pure functions without triggering build-tokens.ts module-level
I/O (readFileSync/writeFileSync). Also renames misleading test.
